requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
passlib[bcrypt]
pyjwt
requests
pillow
google-generativeai